+++ workspace.code-workspace
@@ -1,0 +1,4 @@
+{
+	"folders": [],
+	"settings": {}
+}
